cli: move cli/daemon/internal/appfile to cli/internal/appfile
In preparation for moving logs from being proxied through the
daemon into being handled directly by the client.
